Codec.Crypto.RSA.Pure

Documentation

data RSAError

Constructors

RSAError String

RSAKeySizeTooSmall

RSAIntegerTooLargeToPack

RSAMessageRepOutOfRange

RSACipherRepOutOfRange

RSAMessageTooShort

RSAMessageTooLong

RSAMaskTooLong

RSAIncorrectSigSize

RSAIncorrectMsgSize

RSADecryptionError

RSAGenError GenError

data HashInfo

Constructors

HashInfo

Fields

algorithmIdent :: ByteString

The ASN.1 DER encoding of the hash function identifier.

hashFunction :: ByteString -> ByteString

The hash function

Keys and key generation

data PrivateKey

Represent a RSA private key.

Only the pub, d fields are mandatory to fill.

p, q, dP, dQ, qinv are by-product during RSA generation, but are useful to record here to speed up massively the decrypt and sign operation.

implementations can leave optional fields to 0.

Constructors

PrivateKey

Fields

private_pub :: PublicKey

public part of a private key (size, n and e)

private_d :: Integer

private exponant d

private_p :: Integer

p prime number

private_q :: Integer

q prime number

private_dP :: Integer

d mod (p-1)

private_dQ :: Integer

d mod (q-1)

private_qinv :: Integer

q^(-1) mod p

data PublicKey

Represent a RSA public key

Constructors

PublicKey

Fields

public_size :: Int

size of key in bytes

public_n :: Integer

public p*q

public_e :: Integer

public exponant e

generateKeyPair :: CryptoRandomGen g => g -> Int -> Either RSAError (PublicKey, PrivateKey, g)

Randomly generate a key pair of the given modules length (in bits) to use in any of the following functions. Use of a good random number generator is of considerable importance when using this function. The input CryptoRandomGen should never be used again for any other purpose; either use the output'd generator or throw it all away.

High-level encryption and signature functions

encrypt :: CryptoRandomGen g => g -> PublicKey -> ByteString -> Either RSAError (ByteString, g)

Encrypt an arbitrarily-sized message given the public key and reasonable options. This is equivalent to calling encryptOAEP with SHA-256 as the hash function, MGF1(SHA-256) as the mask generation function, and no label. NOTE: This hash choice means that your key size must be 1024 bits or larger.

encryptOAEP

Arguments

:: CryptoRandomGen g

=> g

-> (ByteString -> ByteString)
The hash function to use

-> MGF
The mask generation function to use

-> ByteString
An optional label to include

-> PublicKey
The public key to encrypt with

-> ByteString
The message to encrypt

-> Either RSAError (ByteString, g)

Encrypt an arbitrarily-sized message using OAEP encoding. This is the encouraged encoding for doing RSA encryption. Note that your key size must be greater than (2 * hash length + 2) * 8. (For example, the encrypt convenience function uses a 256 bit / 32 byte hash function. Thus, its key must be greater than (2 * 32 + 2) * 8 = 528 bits long, and we suggest 1024 as a lower bound.)

encryptPKCS :: CryptoRandomGen g => g -> PublicKey -> ByteString -> Either RSAError (ByteString, g)

Encrypt an arbitrarily-sized message using PKCS1 v1.5 encoding. This encoding is deprecated, and should only be used when interacting with legacy software that cannot be modified.

decrypt :: PrivateKey -> ByteString -> Either RSAError ByteString

Decrypt an arbitrarily-sized message given the public key and reasonable options. This is equivalent to calling encryptOAEP with SHA-256 as the hash function, MGF1(SHA-256) as the mask generation function, and no label.

decryptOAEP

Arguments

:: (ByteString -> ByteString)
The hash function to use

-> MGF
The mask generation function to use

-> ByteString
An optional label to include

-> PrivateKey
The public key to encrypt with

-> ByteString
The message to decrypt

-> Either RSAError ByteString

Decrypt an arbitrarily-sized message using OAEP encoding. This is the encouraged encoding for doing RSA encryption.

decryptPKCS :: PrivateKey -> ByteString -> Either RSAError ByteString

Decrypt an arbitrarily-sized message using PKCS1 v1.5 encoding. This encoding is deprecated, and should only be used when interacting with legacy software that cannot be modified.

sign :: PrivateKey -> ByteString -> Either RSAError ByteString

Compute a signature for the given ByteString, using the SHA256 algorithm in the computation. This is currently defined as rsassa_pkcs1_v1_5_sign hashSHA256. If you want to use a different function, simply use the PKCS function, below; it will accept arbitrarily-length messages.

verify

Arguments

:: PublicKey
The key of the signer

-> ByteString
The message

-> ByteString
The purported signature

-> Either RSAError Bool

Verify a signature for the given ByteString, using the SHA25 algorithm in the computation. Again, if you'd like to use a different algorithm, use the rsassa_pkcs1_v1_5_verify function.

Core routines for OAEP

type MGF = ByteString -> Int64 -> Either RSAError ByteString

A 'mask generation function'. The input is a bytestring, and the output is a hash of the given length. Unless you know what you're doing, you should probably use a MGF1 formulation created with generate_MGF1.

generateMGF1 :: (ByteString -> ByteString) -> MGF

Generate a mask generation function for the rsaes_oaep_*. As suggested by the name, the generated function is an instance of the MGF1 function. The arguments are the underlying hash function to use and the size of a hash in bytes.

The bytestring passed to the generated function cannot be longer than 2^32 * hLen, where hLen is the passed length of the hash.

rsaes_oaep_encrypt

Arguments

:: CryptoRandomGen g

=> g

-> (ByteString -> ByteString)
The hash function to use

-> MGF
An appropriate mask genereation function

-> PublicKey
The recipient's public key

-> ByteString
A label to associate with the message (feel free to use BS.empty)

-> ByteString
The message to encrypt

-> Either RSAError (ByteString, g)

The generalized implementation of RSAES-OAEP-ENCRYPT. Using the default instantiontion of this, provided by the encrypt function, is a pretty good plan if this makes no sense to you, as it is instantiated with reasonable defaults.

The message to be encrypted may not be longer then (k - 2*hLen - 2), where k is the length of the RSA modulus in bytes and hLen is the length of a hash in bytes. Passing in a larger message will generate an error, represented by the Left constructor. Note that this means that OAEP encryption cannot be used with keys smaller than 512 bits.

I have not put in a check for the length of the label, because I don't expect you to use more than 2^32 bytes. So don't make me regret that, eh?

rsaes_oaep_decrypt

Arguments

:: (ByteString -> ByteString)
The hash function to use

-> MGF
A mask generation function

-> PrivateKey
The private key to use

-> ByteString
An optional label whose association with the message should be verified.

-> ByteString
The ciphertext to decrypt

-> Either RSAError ByteString

The generalized implementation of RSAES-OAEP-DECRYPT. Again, decrypt initializes this with a pretty good set of defaults if you don't understand what all of the arguments involve.

The ciphertext message passed to this function must be k bytes long, where k is the size of the modulus in bytes. If it is not, this function will generate an error, represented by the Left constructor.

Futher, k (the length of the ciphertext in bytes) must be greater than or equal to (2 * hLen + 2), where hLen is the length of the output of the hash function in bytes. If this equation does not hold, a (different) error will be generated.

Finally, there are any number of internal situations that may generate an error indicating that decryption failed.

Core PSS routines

|RSASSA-PSS-Sign, RSASSA-PSS-Verify, and the related functions are not included because they are covered by U.S. Patent 7036014, and it's not clear what the restrictions on implementation are. Sorry.

rsaes_pkcs1_v1_5_encrypt :: CryptoRandomGen g => g -> PublicKey -> ByteString -> Either RSAError (ByteString, g)

Implements RSAES-PKCS1-v1.5-Encrypt, for completeness and backward compatibility. Also because I've already written everything else, so why not?

This encryption / padding mechanism has several known attacks, which are described in the literature. So unless you absolutely need to use this for some historical reason, you should avoid it.

The message to be encrypted must be less then or equal to (k - 11) bytes long, where k is the length of the key modulus in bytes.

Because this function uses an unknown amount of randomly-generated data, it takes an instance of RandomGen rather than taking a random number as input, and returns the resultant generator as output. You should take care that you (a) do not reuse the input generator, thus losing important randomness, and (b) choose a decent instance of RandomGen for passing to this function.

rsaes_pkcs1_v1_5_decrypt :: PrivateKey -> ByteString -> Either RSAError ByteString

Implements RSAES-PKCS1-v1.5-Decrypt, for completeness and possible backward compatibility. Please see the notes for rsaes_pkcs_v1_5_encrypt regarding use of this function in new applications without backwards compatibility requirements.

The ciphertext message passed to this function must be of length k, where k is the length of the key modulus in bytes.

rsassa_pkcs1_v1_5_sign

Arguments

:: HashInfo
The hash function to use

-> PrivateKey
The private key to sign with

-> ByteString
The message to sign

-> Either RSAError ByteString
The signature

Generate a signature for the given message using the given private key, using the RSASSA-PKCS1-v1.5-Sign algorithm. Note that in researching the requirements for this project, several independent sources suggested not using the same key across signvalidate and encryptdecrypt contexts. You've been warned.

The output of this function is the signature only, not the message and the signature.

SIZE CONSTRAINT: The size of the public key (in bytes) must be greater than or equal to the length of the hash identifier plus the length of a hash plus 1. Thus, for example, you cannot use a 256 bit RSA key with MD5: 32 (the size of a 256-bit RSA key in bytes) is less than 18 (the size of MD5's identier) + 16 (the size of an MD5 hash in bytes) + 1, or 35.

Thus, * for SHA1 and SHA256, use 512+ bit keys * for SHA384 and SHA512, use 1024+ bit keys

rsassa_pkcs1_v1_5_verify

Arguments

:: HashInfo
The hash function to use

-> PublicKey
The public key to validate against

-> ByteString
The message that was signed

-> ByteString
The purported signature

-> Either RSAError Bool

Validate a signature for the given message using the given public key. The signature must be exactly k bytes long, where k is the size of the RSA modulus IN BYTES.

Hashing algorithm declarations for use in RSA functions

hashSHA1 :: HashInfo

hashSHA224 :: HashInfo

hashSHA256 :: HashInfo

hashSHA384 :: HashInfo

hashSHA512 :: HashInfo

Other mathematical functions that are handy for implementing

largeRandomPrime :: CryptoRandomGen g => g -> Int -> Either RSAError (Integer, g)

Generate a large random prime of a given length in bytes.

generatePQ :: CryptoRandomGen g => g -> Int -> Either RSAError (Integer, Integer, g)

chunkify :: ByteString -> Int64 -> [ByteString]

os2ip :: ByteString -> Integer

i2osp :: Integral a => a -> Int -> Either RSAError ByteString

rsa_dp :: Integer -> Integer -> Integer -> Either RSAError Integer

rsa_ep :: Integer -> Integer -> Integer -> Either RSAError Integer

rsa_vp1 :: Integer -> Integer -> Integer -> Either RSAError Integer

rsa_sp1 :: Integer -> Integer -> Integer -> Either RSAError Integer

modular_inverse

Arguments

:: Integer
e

-> Integer
phi

-> Integer

Compute the modular inverse (d = e^-1 mod phi) via the extended euclidean algorithm.

modular_exponentiation :: Integer -> Integer -> Integer -> Integer

Computes a^b mod c using a moderately good algorithm.

randomBS :: CryptoRandomGen g => g -> Int -> Either RSAError (ByteString, g)

Generate a random ByteString of the given length

randomNZBS :: CryptoRandomGen g => g -> Int -> Either RSAError (ByteString, g)

Create a random bytestring of non-zero bytes of the given length.

Orphan instances

Show SystemRandom

Instance details

Methods

showsPrec :: Int -> SystemRandom -> ShowS

show :: SystemRandom -> String

showList :: [SystemRandom] -> ShowS

Binary PublicKey

Instance details

Methods

put :: PublicKey -> Put

get :: Get PublicKey

putList :: [PublicKey] -> Put

Binary PrivateKey

Instance details

Methods

put :: PrivateKey -> Put

get :: Get PrivateKey

putList :: [PrivateKey] -> Put